Set RHOME env-var to account for Rscript
Rscript is not portable because it uses a compile time absolute path for the R home directory. So any R code that calls something like `system('/path/to/Rscript', ...)` will fail because the Rscript file will in turn call R at the original hard-coded absolute path. For example, parallel::newPSOCKnode() uses this type of command, meaning that any code using PSOCK parallel cluster will not work.

I am not sure if shiny or any of this electron setup ever calls code that in turn makes a call to Rscript. If not, then it is probably fine.

Interestingly, Rscript does check to see if the R home directory is being overridden in the environment, and uses that path instead of the compile time path. However, it does so by looking for the deprecated RHOME env-var instead of R_HOME. Instead of re-compiling Rscript (and wrangling with all the tooling involved in that), we can work around it by assigning into the RHOME env-var as well.